# docs/Enterprise_notebooklm_openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise NotebookLM API
  version: "1.0.0"
  description: >
    API contract cho hệ thống Enterprise NotebookLM (LLM + RAG).
    Bao gồm Workspace Management, Knowledge Base/Document Versioning (Document Understanding qua LlamaParse
    + Hierarchical Chunking), Intelligent Search, AI Chat với Citation (Confidence Engine + Event-driven
    Micro Agents: Rewrite/Graph/SQL), AI Summary, Information Extraction, Multi-document Comparison,
    Report Generation, và Observability.
    Ánh xạ FR1–FR14 (Business_Context.md) và schema v3 (database-design-enterprise-notebooklm.md).
servers:
  - url: https://api.notebooklm.example.com/v1
    description: Production
  - url: https://staging-api.notebooklm.example.com/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: Danh sách Workspace mà user hiện tại có quyền truy cập
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceListResponse" }
    post:
      tags: [Workspaces]
      summary: Tạo Workspace mới (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
  /workspaces/{workspaceId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    get:
      tags: [Workspaces]
      summary: Chi tiết Workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
    patch:
      tags: [Workspaces]
      summary: Cập nhật Workspace (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "200":
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
    delete:
      tags: [Workspaces]
      summary: Xoá Workspace (Admin)
      responses:
        "204": { description: Đã xoá }
...
  /workspaces/{workspaceId}/documents:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    get:
      tags: [Documents]
      summary: Danh sách tài liệu trong Workspace
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - name: file_type
          in: query
          schema: { type: string, enum: [pdf, docx, xlsx, pptx, txt] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentListResponse" }
    post:
      tags: [Documents]
      summary: Upload tài liệu mới (tạo document + version 1, enqueue pipeline)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
      responses:
        "202":
          description: Đã nhận, pipeline đang xử lý bất đồng bộ
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentVersion" }
  /workspaces/{workspaceId}/documents/{documentId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      tags: [Documents]
      summary: Chi tiết tài liệu (kèm current_version)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
    delete:
      tags: [Documents]
      summary: Xoá tài liệu (mọi version)
      responses:
        "204": { description: Đã xoá }
  /workspaces/{workspaceId}/documents/{documentId}/chunks:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      tags: [Documents]
      summary: Danh sách chunk bản hiện hành (Document Viewer / deep-link Search)
      description: >
        Trả chunks đã index của current_version (hoặc versionId) để viewer
        scroll/highlight theo chunk_id. Không gọi Hybrid Retrieval.
      parameters:
        - name: versionId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentChunkListResponse" }
        "404":
          description: Document không tồn tại
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /workspaces/{workspaceId}/documents/{documentId}/content:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      tags: [Documents]
      summary: Stream file gốc (hoặc preview PDF) cho Original Document Viewer
      description: >
        Trả bytes Original Representation từ MinIO. Không bao giờ trả markdown.
        PDF gốc (hoặc document.pdf sibling) dùng để render; download=true để tải về.
      parameters:
        - name: versionId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: download
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: >
            true = tải Original file (DOCX/PDF gốc). false = Preview Representation
            (chỉ khi preview_status=completed; ngược lại 409).
      responses:
        "200":
          description: Binary content (application/pdf hoặc MIME gốc)
          content:
            application/pdf:
              schema: { type: string, format: binary }
            application/octet-stream:
              schema: { type: string, format: binary }
        "409":
          description: Preview chưa sẵn sàng (pending/processing/failed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Document/version không tồn tại
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /workspaces/{workspaceId}/documents/{documentId}/versions:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      tags: [Documents]
      summary: Lịch sử version của tài liệu
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DocumentVersion" }
    post:
      tags: [Documents]
      summary: Upload lại (thay thế nội dung) → tạo version mới, enqueue pipeline
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "202":
          description: Version mới đang xử lý
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentVersion" }
  /workspaces/{workspaceId}/documents/{documentId}/versions/{versionId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
      - { $ref: "#/components/parameters/VersionId" }
    get:
      tags: [Documents]
      summary: Chi tiết 1 version (status, checksum, page_count...)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentVersion" }
  /workspaces/{workspaceId}/documents/{documentId}/versions/{versionId}/set-current:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
      - { $ref: "#/components/parameters/VersionId" }
    post:
      tags: [Documents]
      summary: Đặt version này làm current_version_id (rollback/chuyển version)
      responses:
        "200":
          description: Đã cập nhật current version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
  /workspaces/{workspaceId}/documents/{documentId}/versions/{versionId}/pipeline-status:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/DocumentId" }
      - { $ref: "#/components/parameters/VersionId" }
    get:
      tags: [Documents]
      summary: Trạng thái pipeline_run + từng pipeline_stage_log (Document Understanding/Cleaning/Hierarchical Chunking/Embedding/Graph/Indexing)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineRun" }

  # ---------------------------------------------------------------------
  # SEARCH (FR3, Module 3)
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VersionId:
      name: versionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        title: { type: string }
        file_type: { type: string, enum: [pdf, docx, xlsx, pptx, txt] }
        current_version_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DocumentListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Document" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    DocumentVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        uploaded_by: { type: string, format: uuid }
        version_number: { type: integer }
        file_size_bytes: { type: integer }
        checksum_sha256: { type: string }
        page_count: { type: integer }
        status: { type: string, enum: [processing, ready, failed] }
        is_current: { type: boolean }
        parser: { type: string, example: llamaparse, description: "Engine dùng cho Document Understanding" }
        created_at: { type: string, format: date-time }
    PipelineRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        document_version_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, running, completed, failed] }
        retry_count: { type: integer }
        error_message: { type: string, nullable: true }
        stages:
          type: array
          items: { $ref: "#/components/schemas/PipelineStageLog" }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
    PipelineStageLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        stage:
          type: string
          enum: [document_understanding, cleaning_normalize, hierarchical_chunking, embedding, graph_extraction, indexing]
        status: { type: string, enum: [pending, running, completed, failed] }
        duration_ms: { type: integer }
        metadata: { type: object }
        error_message: { type: string, nullable: true }
...
    DocumentChunkListResponse:
      type: object
      properties:
        document_id: { type: string, format: uuid }
        document_version_id: { type: string, format: uuid, nullable: true }
        document_title: { type: string }
        file_type: { type: string, enum: [pdf, docx, xlsx, pptx, txt] }
        viewer_kind: { type: string, enum: [pdf, original_download] }
        preview_status:
          type: string
          enum: [pending, processing, completed, failed]
        preview_type:
          type: string
          enum: [pdf, html, image]
          nullable: true
        preview_generated_at: { type: string, format: date-time, nullable: true }
        heading_tree:
          type: array
          items: { type: object }
        items:
          type: array
          items: { $ref: "#/components/schemas/DocumentChunk" }
    DocumentChunk:
      type: object
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        document_version_id: { type: string, format: uuid }
        chunk_index: { type: integer }
        content: { type: string }
        page_number: { type: integer, nullable: true }
        section_index: { type: integer, nullable: true }
        section: { type: string, nullable: true }
        heading_path: { type: string, nullable: true }
        section_path: { type: string, nullable: true }
        bounding_box:
          type: array
          items: { type: number }
          nullable: true
          description: Optional [x, y, w, h] from layout_metadata
        start_offset: { type: integer, nullable: true }
        end_offset: { type: integer, nullable: true }
...